* Added JavaScript roman numerals code kata.

+++ julia/PrimeNumbers/src/PrimeNumbers.jl
@@ -8,13 +8,13 @@ export find
 #
 # @brief Computes the nth prime number.
 #
-# @param n The prime number position estarting from 1.
+# @param n The prime number position starting from 1.
 #
 # @return The nth prime integer.
 #
 function compute(n::Integer)
-  known_primes = Int64[]
-  position     = 2
+  known_primes    = Int64[]
+  position::Int64 = 2
 
   while length(known_primes) < n
     if findfirst(x -> (rem(position, x) == 0), known_primes) == 0
